chore: add root tests/conftest.py with shared fixtures
Move shared test fixtures (sample_dataset, empty_dataset, temp_dir,
MockBlock, mock_block) from tests/flow/conftest.py to a new root
tests/conftest.py so they are available to all test modules without
duplication.

Update MockBlock import paths in test_agent_config.py, test_base.py,
and test_time_estimation.py. Remove resolved items from
tech-debt-tracker.md.
