+++ docs/agent-knowledge/tech-debt-tracker.md
@@ -5,28 +5,11 @@ resolved; new items are added as they are discovered.
 
 ## High Priority
 
-- **Legacy .isort.cfg and .pylintrc redundant with ruff.** These config files
-  are superseded by ruff which handles both linting and import sorting. Being
-  removed in Phase 1.
-
-- **Stale \_\_pycache\_\_ directories for deleted test modules.** Leftover
-  bytecode caches exist for test modules that have been removed:
-  `tests/blocks/evaluation/`, `tests/blocks/utilblocks/`,
-  `tests/blocks/column_ops/`. These should be cleaned up and added to
-  `.gitignore` if not already excluded.
-
 - **Two documentation systems (MkDocs + Next.js) creating drift risk.** Having
   two separate documentation builds means content can diverge. Consolidate to
   a single system or establish a clear ownership boundary between them.
 
 ## Medium Priority
-
-- **No top-level conftest.py with shared test fixtures.** Common test helpers
-  (mock LLM clients, sample datasets, temporary flow YAML) are duplicated
-  across test modules instead of shared via a root `conftest.py`.
-
-- **No .env.example at project root.** New contributors have no reference for
-  which environment variables are expected. Being added in Phase 1.
 
 - **Some block categories missing in CLAUDE.md block category table.** The
   `blocks/code` category and any newly added categories are not listed in the
